fixing doc build script in this repo to prevent checkout errors

+++ doc_build.sh
@@ -1,16 +1,21 @@
-# Minimal makefile for Sphinx documentation
-#
-BRANCH_NAME=$1
+#!/bin/bash
 
-if [ -z ${BRANCH_NAME+x} ] ;
-  then echo "please provide a branch name from which documentation will be
-    built";
-  else echo "building docs from $BRANCH_NAME"
+if [ $# -ne 1 ]; then
+  echo "Error: Please provide a branch name from which documentation will be built";
+  exit 1
 fi
 
-echo "Building documentation"
+BRANCH_NAME=$1
+
+echo "Building documentation from $BRANCH_NAME"
 echo "checking out gh-pages"
-git checkout gh-pages
+if ! git checkout gh-pages
+then
+  echo >&2 "checkout of gh-pages branch failed; please ensure you have local changes commited prior to running this script "
+  echo "exiting"
+  exit 1
+fi
+
 pwd
 echo "removing current files"
 rm -r *.html *.js docs/
@@ -20,7 +25,10 @@ mv docs/* .
 make html
 mv -fv build/html/* ./
 rm -r tweet_parser docs build Makefile source
-git add -A
-git commit -m "Generated gh-pages for `git log $BRANCH_NAME -1 --pretty=short \ --abbrev-commit`"
-# git push origin gh-pages
-# git checkout $BRANCH_NAME
+echo "--------------------------------------------------------"
+echo "docs built; please review these changes and then run the following:"
+echo "--------------------------------------------------------"
+echo git add -A
+echo git commit -m "Generated gh-pages for `git log $BRANCH_NAME -1 --pretty=short --abbrev-commit`"
+echo git push origin gh-pages
+echo git checkout $BRANCH_NAME
